+++ tests/manual/run_input_generator.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+topic="sentinel/collect/generator$$"
+socket="ipc:///tmp/proxy_input.sock"
+
+./input_generator.py \
+	--socket $socket \
+	--topic $topic